Brought back reactivity *and* finally fixed the issue with items jumping to their previous position after being dragged.
Removed autopublish ands explicitly published "items" and "boxes".

The problem with the jump back was *not* a problem with reactivity. It was that when using d3 to sync the properties of the child rects of the box and item <g>s, the updated data was not set in those child rects, so d3 would set those x and y attributes to the last value stored with the rect elements.

Passing a function that returns the object that was passed to it in the .data() function takes care of making sure data is set recursively.

Added a makeSureItemsAreInFrontOfBoxes method to deal with unpredictable loading times and order between items and boxes.

Set up explicit dependency contexts for redrawing boxes and items and called those functions during the subscription handlers for boxes and items, respectively. Dropped autorun during debugging. Leaving it out for now.