Fields are now editable in place. Doing this by bringing in makeEditable(), which adds a <foreignObject> to a clicked svg element, then deletes it when it's no longer needed.
To allow the handlers added by makeEditable to get called, needed to make saveAndRecalcOnDragStop *not* save if the drag did not actually change anything. To keep track of drag changes, added the Dragger object with its currentDragChangeX and currentDragChangeY members and methods that update and reset them. Moved the drag handler functions into that object.

Also added syncDataToCollection to handle updating various fields in a record using a datum.