Can now create and use multiple boards. Privacy stuff is not added yet.
Rearranged collection subscriptions so that:
- They're no longer nested.
- They're inside of an autorun(), which doesn't seem to create problems now.
- They also reference the currentBoard session var so that changes to that trigger their callbacks.

Got rid of 0-array guards for syncNodes* methods. It's important to call them when there's no data so that nodes get deleted to match the data.

Fixed a .classed call that forgot the value param. Without that, it tries to fetch the class, which can make it crash if the selection is null.

Implemented save event for new board dialog template that calls createBoard.

Added boardList template that sets the currentBoard session variable when clicked.

Models: Added createBoard method. Added board param to existing create methods.