Got text to be editable at other zoom levels by zooming to 1 and centering on the editable item.
Pass lockZoomToDefaultCenterPanAtDataCoords to makeEditable as doOnStartEdit. lockZoomToDefaultCenterPanAtDataCoords centers the pan on the element for the given data object and sets the zoom to 1.

editableSVGText: Use the d object in doOnEditStart and the exitEditing call in the click handler. (That call to exitEditing was being passed garbage before and was working by accident.)